requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.24.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-colorfield==0.12.0
django-js-asset==2.2.0
django-nested-admin==4.1.1
django-simple-menu==2.1.3
django-sitetree==1.18.0
django-storages==1.14.4
django_csp==3.8
et_xmlfile==2.0.0
idna==3.10
isodate==0.7.2
mysqlclient==2.2.4
numpy
openpyxl==3.1.5
pandas
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
XlsxWriter==3.2.2